remodularize storage module
