+++ src/storage/mod.rs
@@ -1,5 +1,14 @@
-//! Implementation of functionality related to the management of data in between persistent /
-//! non-volatile storage and volatile memory.
+//! This module contains the definition and implementation of [`Frame`] and [`FrameGroup`], which
+//! are types that represent the buffer frames that the buffer pool manager is in charge of.
+//!
+//! A [`Frame`] is intended to hold [`PAGE_SIZE`] bytes of data, and is also intended to be shared
+//! with the the kernel to avoid unnecessary `memcpy`s from the kernel's internal buffers into
+//! user-space buffers.
+//!
+//! A [`FrameGroup`] instance groups [`Frame`]s together so that evictions do not have to search
+//! every single [`Frame`] in the buffer pool for an eviction candidate.
 
+pub(crate) mod eviction;
 pub(crate) mod frame;
+pub(crate) mod frame_group;
 pub(crate) mod storage_manager;
